# src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Backend API",
    "description": "Production-ready backend API for an advanced e-commerce platform.",
    "version": "1.0.0",
    "contact": {
      "name": "Your Name",
      "email": "youremail@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["user", "admin"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "category": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": { "type": "string" },
                "quantity": { "type": "integer" }
              }
            }
          },
          "total": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "paid", "shipped", "delivered", "cancelled"] },
          "paymentId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "productId": { "type": "string" },
          "userId": { "type": "string" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Delivery": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "orderId": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "in_transit", "delivered"] },
          "trackingNumber": { "type": "string" },
          "carrier": { "type": "string" },
          "estimatedDelivery": { "type": "string", "format": "date-time" }
        }
      },
      "Promotion": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "discount": { "type": "number" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "active": { "type": "boolean" }
        }
      },
      "Wishlist": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "products": { "type": "array", "items": { "type": "string" } }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/User" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT token" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/products": {
      "get": {
        "summary": "Get all products",
        "tags": ["Products"],
        "responses": { "200": { "description": "List of products" } }
      },
      "post": {
        "summary": "Create a product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": { "201": { "description": "Product created successfully" } }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "tags": ["Products"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product details" } }
      },
      "put": {
        "summary": "Update product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
        },
        "responses": { "200": { "description": "Product updated" } }
      },
      "delete": {
        "summary": "Delete product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Product deleted" } }
      }
    },
    "/api/orders": {
      "get": {
        "summary": "Get all orders",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of orders" } }
      },
      "post": {
        "summary": "Create a new order",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
        },
        "responses": { "201": { "description": "Order created successfully" } }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order by ID",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Order details" } }
      },
      "put": {
        "summary": "Update order status",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
        "responses": { "200": { "description": "Order updated" } }
      }
    }
  }
}